requirements.txt
flask
py-cpuinfo
psutil
gpustat
